requirements.txt
azure-functions==1.20.0
openai>=1.40.0
tenacity>=8.4.2
